refactor: add esmExternals to build options
